refactor(cleanup): split animation module, merge binary into data, move column_batch to raycaster
- Split src/animation/mod.rs (590 lines) into frame.rs, clip.rs, event.rs, controller.rs
- Delete src/doll/ (empty, no Rust code)
- Merge src/binary/ into src/data/ via new data/bin_pack.rs; expose write/read/size on luna.data
- Move graphics/column_batch.rs to raycaster/column_batch.rs (only uses crate::math::Color)
- Update sprite_api.rs import; update test imports for graphics_ext_tests and data_tests
- Fix clippy: module-inception in animation, redundant closures in serial/csv, unnecessary casts in raycaster/minimap_overlay

+++ src/raycaster/minimap_overlay.rs
@@ -76,8 +76,8 @@ pub fn extract_minimap(
     }
 
     // Draw player dot at center
-    let center_px = (view_radius * cell_size + cell_size / 2) as u32;
-    let center_py = (view_radius * cell_size + cell_size / 2) as u32;
+    let center_px = view_radius * cell_size + cell_size / 2;
+    let center_py = view_radius * cell_size + cell_size / 2;
     draw_player_arrow(
         &mut pixels,
         pixel_w,
